Pull NestedList out of ArchiveTree
